+++ src/__tests__/perf-marks-without-user-timing.ts
@@ -1,21 +1,30 @@
 jest.doMock('../is-user-timing-api-supported', () => ({
   isUserTimingAPISupported: false,
 }));
+jest.doMock('../is-performance-observable-supported', () => ({
+  isPerformanceObservableSupported: false,
+}));
 
 import * as PerfMarks from '../marks';
 
 describe('PerfMarks: User timing API is NOT available', () => {
   beforeEach(() => {
     spyOn(Date, 'now').and.callThrough();
+    spyOn(performance, 'now').and.callThrough();
+    spyOn(performance, 'mark');
+    spyOn(performance, 'clearMeasures');
+    spyOn(performance, 'clearMarks');
   });
 
   afterEach(() => {
     jest.clearAllMocks();
   });
 
-  it('should use `performance.mark` if user calls start mark', () => {
+  it('should store datetime in memory if user calls start mark', () => {
     const mark = 'mark';
+
     PerfMarks.start(mark);
+    expect(performance.now).not.toHaveBeenCalled();
     expect(Date.now).toHaveBeenCalled();
   });
 
@@ -23,6 +32,12 @@ describe('PerfMarks: User timing API is NOT available', () => {
     const mark = 'mark';
     PerfMarks.start(mark);
     PerfMarks.clear(mark);
+
+    expect(performance.now).not.toHaveBeenCalled();
+    expect(Date.now).toHaveBeenCalled();
+    expect(performance.clearMeasures).not.toHaveBeenCalled();
+    expect(performance.clearMarks).not.toHaveBeenCalledWith();
+    expect(performance.getEntriesByName(mark)).toHaveLength(0);
   });
 
   it('should remove all markers and measures if `clearAll` is called', () => {
@@ -30,6 +45,12 @@ describe('PerfMarks: User timing API is NOT available', () => {
     PerfMarks.start('mark-2');
 
     PerfMarks.clearAll();
+
+    expect(performance.now).not.toHaveBeenCalled();
+    expect(Date.now).toHaveBeenCalled();
+    expect(performance.clearMeasures).not.toHaveBeenCalled();
+    expect(performance.clearMarks).not.toHaveBeenCalled();
+    expect(performance.getEntries()).toHaveLength(0);
   });
 
   it('should return user timing information if user finishes mark', () => {
@@ -43,5 +64,9 @@ describe('PerfMarks: User timing API is NOT available', () => {
         startTime: expect.any(Number),
       }),
     );
+  });
+
+  it('should return `isPerformanceObservableSupported` value as `false` if PerformanceObservable API is NOT available', () => {
+    expect(PerfMarks.isPerformanceObservableSupported).toBeFalsy();
   });
 });