+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './marks';
 export * from './entries';
 export * from './is-user-timing-api-supported';
+export * from './is-performance-observable-supported';